Add a custom activation script to link app bundles on Darwin
macOS's desktop doesn't like the way nix-darwin/home-manager symlink the
app bundles by default. They don't show up in Spotlight and they can't
be pinned to the Dock, both things I absolutely want apps to be able to
do if I am going to have them be installed by Nix.

Playing around a bit I discovered that macOS is fine with symlinking the
`Contents` directory inside the bundle, and will treat a fake bundle
with symlinked `Contents` just like it was the real one, with none of
the downsides of the default approach of symlinking the entire .app.

The way this is set up was inspired by https://github.com/IvarWithoutBones/dotfiles/commit/0b3faad8bd